Login functionality started
Added JWT authorization via cookie
used AuthGuard of Angular
and checked email id and password.

Next plan is to complete AuthGuard with best practices and give prompt to user that invalid password. right now getting in console because main focus of this push is jwt and AuthGuard.

now password is also stored in HASH. that change also done.

and may be i will implement like user can use Username or Email to login